call: add new call flag CTDB_CALL_FLAG_VACUUM_MIGRATION
This is to be used when the CTDB_SRVID_VACUUM_FETCH message
triggers the migration of deleted records to the lmaster.
The lmaster can then delete records that have not been
migrated with data instead of storing them.

similar to commit 455cc6616e10b7f09589f9b87cb60f591bb502b0.

Signed-off-by: Stefan Metzmacher <metze@samba.org>