ctdb_private.h: add record flag CTDB_REC_FLAG_AUTOMATIC
This is a flag that shall signa that a record has been automatically generated by ctdb
and not by an explicit client store operation. This will be used in the ctdb_ltdb_fetch
operation which stores an empty record with default initial header before trying to
migrate the record from the dmaster when the record does not exist in the local tdb.

similar to commit 46381a3cb58ccc11422af8f7798c80ea8d72294f.

Signed-off-by: Stefan Metzmacher <metze@samba.org>