ctdb_ltdb_store_server: honour the AUTOMATIC record flag
Do not delete empty records that carry this flag but store
them and schedule them for deletetion. Do not store the flag
in the ltdb though, since this is internal only and should not
be visible to the client.
(cherry picked from commit f898ff21fa338358179e79381215b13a6bc77c53)

Signed-off-by: Stefan Metzmacher <metze@samba.org>