CTDB_FATAL() try to drop all ip addresses when shutting down through ctdb_fatal()
If we shutdown through a ctdb_fatal(), unless there is an external mechanism to trap this and restart ctdb, we will get a duplicate address problem since addresses held on the node remain which collides with the addresses when failed over to other nodes.

During ctdb_fatal(), spawn a child process to try to drop all held public addresses.